clean out more desktop reducers
* move notification system into core app reducer
* delete unused configLastSaved property
